Create EstimatedCallItem component
